Refactor(dataSync): Some clean ups around the dataSync.
