+++ app/src/main/graphql/com.m.dataSync/addTaskSubscription.graphql
