+++ app/src/main/graphql/com.m.dataSync/deleteTask.graphql
