+++ app/src/main/graphql/com.m.dataSync/deleteTaskSubscription.graphql
