+++ app/src/main/graphql/com.m.dataSync/fragment.graphql
