+++ app/src/main/graphql/com.m.dataSync/getAllTasks.graphql
