+++ app/src/main/graphql/com.m.dataSync/schema.json
