+++ app/src/main/java/com/m/dataSync/CreateTask.java
@@ -1,4 +1,4 @@
-package com.m.helper;
+package com.m.dataSync;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -7,25 +7,25 @@ import android.view.View;
 import android.widget.EditText;
 
 import com.apollographql.apollo.ApolloCall;
+import com.apollographql.apollo.ApolloClient;
 import com.apollographql.apollo.api.Response;
 import com.apollographql.apollo.exception.ApolloException;
-import com.m.androidNativeApp.CreateTaskMutation;
 import com.m.androidNativeApp.MainActivity;
 import com.m.androidNativeApp.R;
-import com.m.services.appAuth.AppAuthActivity;
+import com.m.helper.LoginActivity;
 
 import org.jetbrains.annotations.NotNull;
 
-import static com.m.androidNativeApp.MainActivity.client;
-import static com.m.services.appAuth.AppAuthActivity.RE_AUTH;
+import static com.m.helper.LoginActivity.RE_AUTH;
 
 
 public class CreateTask extends Activity {
-
+    private ApolloClient client;
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
+        client = Client.getInstance();
         setContentView(R.layout.add_task);
+
     }
 
     /**
