+++ app/src/main/java/com/m/dataSync/ItemAdapter.java
@@ -1,24 +1,23 @@
-package com.m.helper;
+package com.m.dataSync;
 
 import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
-import android.widget.Button;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.m.androidNativeApp.R;
+import com.m.helper.Item;
 
 import java.util.List;
 
-public class ItemAdapter extends RecyclerView.Adapter<ItemAdapter.ItemViewHolder> {
+public class ItemAdapter extends RecyclerView.Adapter<ItemViewHolder> {
     Context mCtx;
     private List<Item> itemList;
 
-    public ItemAdapter(Context mCtx, List<Item> itemList) {
+    ItemAdapter(Context mCtx, List<Item> itemList) {
         this.mCtx = mCtx;
         this.itemList = itemList;
     }
@@ -47,21 +46,6 @@ public class ItemAdapter extends RecyclerView.Adapter<ItemAdapter.ItemViewHolder
     @Override
     public int getItemCount() {
         return itemList.size();
-    }
-
-    class ItemViewHolder extends RecyclerView.ViewHolder {
-
-        TextView title, description;
-        Button deleteButton;
-
-        public ItemViewHolder(@NonNull View itemView) {
-            super(itemView);
-
-            title = itemView.findViewById(R.id.titleView);
-            description = itemView.findViewById(R.id.descriptionView);
-            deleteButton = itemView.findViewById(R.id.deleteButton);
-
-        }
     }
