+++ app/src/main/res/layout/list_layout.xml
@@ -10,24 +10,24 @@
 
         <TextView
             android:id="@+id/titleView"
-            android:layout_width="match_parent"
+            android:layout_width="275dp"
             android:layout_height="wrap_content"
+            android:layout_marginLeft="10dp"
             android:layout_marginTop="10dp"
             android:fontFamily="monospace"
-            android:text="TextView"
-            android:layout_marginLeft="10dp"
+            android:text="@string/textview"
             android:textColor="#000000"
             android:textSize="18sp"
             android:textStyle="bold" />
 
         <TextView
             android:id="@+id/descriptionView"
-            android:layout_width="match_parent"
+            android:layout_width="267dp"
             android:layout_height="wrap_content"
-            android:layout_marginTop="35dp"
             android:layout_marginLeft="15dp"
+            android:layout_marginTop="35dp"
             android:fontFamily="monospace"
-            android:text="TextView"
+            android:text="@string/textview"
             android:textColor="#000000" />
 
         <Button
@@ -39,7 +39,7 @@
             android:layout_marginTop="10dp"
             android:background="#00F8F7F7"
             android:onClick="deleteTask"
-            android:text="Delete"
+            android:text="@string/delete"
             android:textSize="14sp"
             android:textStyle="bold" />
     </androidx.cardview.widget.CardView>
